+++ lib/modules/storage/workers/prune_trash_job.ex
@@ -13,7 +13,7 @@ defmodule PhoenixKit.Modules.Storage.Workers.PruneTrashJob do
 
   alias PhoenixKit.Modules.Storage
   alias PhoenixKit.Modules.Storage.Libraries
-  alias PhoenixKit.Modules.Storage.Workers.LocationBackfillJob
+  alias PhoenixKit.Modules.Storage.Workers.{ChecksumBackfillJob, LocationBackfillJob}
 
   @impl Oban.Worker
   def perform(_job) do
@@ -21,6 +21,7 @@ defmodule PhoenixKit.Modules.Storage.Workers.PruneTrashJob do
 
     # Records where objects stored before V204 are, while any are left.
     _ = LocationBackfillJob.maybe_enqueue()
+    _ = ChecksumBackfillJob.maybe_enqueue()
 
     case Libraries.queue_expired_purges(days) do
       0 -> :ok